+++ mapbox_maps_flutter/example/lib/platform.dart
@@ -3,3 +3,4 @@ import 'package:flutter/foundation.dart';
 final isAndroid = !kIsWeb && defaultTargetPlatform == TargetPlatform.android;
 final isIOS = !kIsWeb && defaultTargetPlatform == TargetPlatform.iOS;
 final isMobile = isAndroid || isIOS;
+final isWeb = kIsWeb;